Change the creation of the APIContext to a prototype bean and update the usages to use an @Lookup method instead of an injected property
This change was done because the PayPal-Request-Id is only generated once per creation of a Payment. This is an issue because if two customers create payments and then both checkout before a 3rd customer creates a payment then, because it is a singleton bean, both customers will attempt to checkout with the same PayPal-Request-Id which causes a "Duplicate PayPal-Request-Id" exception from PayPal. By using a prototype bean and using a new APIContext for every call to PayPal we can ensure that every request will have it's own unique PayPal-Request-Id